[KOGITO-6267] move PackageClassLoader into dynamic module to fix kogito native compilation
