+++ drools-kiesession/pom.xml
@@ -37,11 +37,6 @@
       <groupId>org.drools</groupId>
       <artifactId>drools-core-reflective</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.drools</groupId>
-      <artifactId>drools-core-dynamic</artifactId>
-      <scope>runtime</scope>
-    </dependency>
 
     <dependency>
       <groupId>org.drools</groupId>
