+++ drools-traits/src/test/java/org/drools/traits/core/factmodel/InstancesHashcodedTest.java
@@ -25,7 +25,7 @@ import org.drools.core.factmodel.ClassBuilder;
 import org.drools.core.factmodel.ClassDefinition;
 import org.drools.core.factmodel.FieldDefinition;
 import org.drools.core.rule.JavaDialectRuntimeData;
-import org.drools.core.rule.JavaDialectRuntimeData.PackageClassLoader;
+import org.drools.dynamic.PackageClassLoader;
 import org.drools.reflective.classloader.ProjectClassLoader;
 import org.junit.Test;
 
@@ -53,7 +53,7 @@ public class InstancesHashcodedTest {
         byte[] d = builder.buildClass( classDef, prjClassLoader );
         JavaDialectRuntimeData data = new JavaDialectRuntimeData();
         data.write( convertClassToResourcePath(classDef.getClassName()), d );
-        ClassLoader classLoader = new PackageClassLoader(data, prjClassLoader);
+        ClassLoader classLoader = new PackageClassLoader(data.getStore(), prjClassLoader);
         
         ClassFieldAccessorStore store = new ClassFieldAccessorStore();
         store.setClassFieldAccessorCache( new ClassFieldAccessorCache( classLoader ) );
